Refactored to reduce sub-fragments
